pyproject.toml
[project]
name = "genimi-serach-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.8.0",
    "python-dotenv>=1.1.0",
]
